[MOB-19] Moving command-handler and fixing imports
